+++ src/page/main/SearchPage.js
@@ -1,3 +1,4 @@
+import { useNavigate } from "react-router-dom";
 import {useContext, useEffect} from "react";
 import {SongContext} from "../../layout/MainLayout";
 import {Box, Flex} from "@chakra-ui/react";
@@ -5,6 +6,12 @@ import SongRequestComp from "../../component/SongRequestComp";
 
 export function SearchPage() {
   const { searched } = useContext(SongContext);
+  const navigate = useNavigate(); // 페이지 이동
+
+  // 노래 상세 페이지로 이동하는 함수
+  const goToSongPage = (songId) => {
+    navigate(`/main/song/${songId}`);
+  };
 
   // 새로고침 방지 변수
   const preventClose = (e:BeforeUnloadEvent) => {
@@ -33,6 +40,9 @@ export function SearchPage() {
             width={"70%"}
             justifyContent={"space-between"}
             border={"1px solid black"}
+            // Box를 클릭하면 해당 노래의 상세 페이지로 이동
+            onClick={() => goToSongPage(song.id)}
+            style={{ cursor: "pointer" }}
           >
             <Box>{song.title}</Box>
             <Box>{song.artistName}</Box>
